+++ jsolex-core/src/main/java/me/champeau/a4j/jsolex/processing/sun/SpectralLineRealigner.java
@@ -109,7 +109,6 @@ public class SpectralLineRealigner {
     private static boolean isUsable(SpectrumAnalyzer.QueryDetails details) {
         return details != null
                && details.line() != null
-               && !SpectralRay.AUTO.equals(details.line())
                && details.line().wavelength().angstroms() > 0
                && details.pixelSize() > 0
                && details.binning() > 0